Moved out IEEE754tools.h from supla-common
