+++ src/supla/tools.h
@@ -20,7 +20,7 @@
 #define _tools_H_
 
 #include <Arduino.h>
-#include "supla-common/IEEE754tools.h"
+#include "supla/IEEE754tools.h"
 
 void float2DoublePacked(float number, uint8_t *bar, int byteOrder = LSBFIRST);
 float doublePacked2float(uint8_t *bar);
